+++ CHANGES.rst
@@ -1,4 +1,14 @@
 
+0.0.7 (?)
+---------
+
+- :code:`cachedjit` decorator
+
+0.0.6 (2018-11-05)
+------------------
+
+- Type annotations to define Pythran functions and blocks
+
 0.0.5 (2018-10-14)
 ------------------